Improve readability, restrict pw to TPUs, fix help.
