+++ xpk.py
@@ -41,6 +41,7 @@ import subprocess
 import sys
 import tempfile
 import time
+import xpk_helpers.pathways_xpk_helpers as pw
 from dataclasses import dataclass
 
 ################### Compatibility Check ###################
@@ -2435,7 +2436,7 @@ def get_main_container(args, system, docker_image, resource_type) -> str:
   """
   return yaml.format(args=args,
                    system=system,
-                   image_pull_policy=add_image_pull_policy_for_pw(args),
+                   image_pull_policy=pw.add_image_pull_policy_for_pw(args),
                    env=get_env_container(args),
                    container_ports=add_container_ports(args),
                    jax_coordinator_port=add_jax_coordinator_port(system),
@@ -2444,21 +2445,7 @@ def get_main_container(args, system, docker_image, resource_type) -> str:
                    command=command,
                    xpk_internal_commands=xpk_internal_commands,
                    resources=get_main_container_resources(args, system, resource_type),
-                   pw_volume_mounts= get_pw_volume_mounts(args))
-
-def add_image_pull_policy_for_pw(args):
-  """ Add image pull policy only for Pathways containers.
-  Args:
-    args: user provided args.
-
-  Returns:
-    str:
-      YAML stating that the image will be pulled fro GCR every time.
-  """
-  yaml="""imagePullPolicy: Always"""
-  if args.use_pathways:
-    return yaml.format(args=args)
-  return ""
+                   pw_volume_mounts=pw.get_pw_volume_mounts(args))
 
 def get_env_container(args):
   """ Environment configuration for the main container.
@@ -2475,25 +2462,14 @@ def get_env_container(args):
                 - name: JAX_PLATFORMS
                   value: proxy
                 - name: JAX_BACKEND_TARGET
-                  value: grpc://{args.workload}-proxy-0-0.{args.workload}:38676"""
+                  value: grpc://{args.workload}-proxy-0-0.{args.workload}:38676
+                - name: JOBSET_NAME
+                  valueFrom:
+                    fieldRef:
+                      fieldPath: metadata.annotations['jobset.sigs.k8s.io/jobset-name']"""
   if args.use_pathways:
     return env_yaml.format(args=args)
   return args.env
-
-def get_pw_volume_mounts(args) -> str:
-  """ Resources for the main container.
-  Args:
-    args: user provided args.
-
-  Returns:
-    str:
-      YAML for the volumes mounted within a Pathways container as a YAML string.
-  """
-  volume_yaml="""- mountPath: /tmp
-                  name: shared-tmp"""
-  if args.use_pathways:
-    return volume_yaml
-  return ""
 
 def get_main_container_resources(args, system, resource_type) -> str:
   """ Resources for the main container.
@@ -2504,7 +2480,7 @@ def get_main_container_resources(args, system, resource_type) -> str:
 
   Returns:
     str:
-      Pathways resources port as a YAML string
+      Workload resources port as a YAML string
   """
   # Resources requirements for Pathways workload containers are known.
   resources_yaml="""cpu: "24"
@@ -2749,73 +2725,6 @@ def get_cpu_affinity(accelerator_type) -> str:
     return yaml
   return ""
 
-def get_pathways_rm_args(args) -> str:
-  """Arguments for the Pathways resource manager.
-  Args:
-    args: user provided arguments for running the command.
-
-  Returns:
-    str: yaml containing arguments for the Pathways resource manager.
-  """
-  yaml="""- --alsologtostderr
-              - --pathways_server_port=38677
-              - --pathways_server_provides_devices=false
-              - --pathways_device_type=NONE
-              - --pathways_persistent_compilation_cache=false
-              - --pathways_compilation_mode=compile_at_worker
-              - --pathways_tmp_dir_pattern={args.pathways_gcs_location}
-              - --pathways_resource_manager_expected_num_worker_jobs={args.num_slices}"""
-  if args.use_pathways:
-    return yaml.format(args=args)
-  else:
-    return ""
-
-def get_pathways_worker_args(args) -> str:
-  """Arguments for the Pathways workers.
-  Args:
-    args: user provided arguments for running the command.
-
-  Returns:
-    str: yaml containing arguments for the Pathways workers.
-  """
-  yaml="""- --alsologtostderr
-              - --pathways_server_port=38677
-              - --pathways_resource_manager={args.workload}-rm-0-0.{args.workload}:38677
-              - --pathways_persistent_compilation_cache=false
-              - --pathways_compilation_mode=compile_at_worker
-              - --xla_tpu_enable_data_parallel_all_reduce_opt=true
-              - --xla_tpu_data_parallel_opt_different_sized_ops=true
-              - --xla_tpu_enable_async_collective_fusion=true
-              - --xla_tpu_enable_async_collective_fusion_fuse_all_gather=true
-              - --xla_tpu_enable_async_collective_fusion_multiple_steps=true
-              - --xla_tpu_overlap_compute_collective_tc=true
-              - --xla_enable_async_all_gather=true
-              - --pathways_tmp_dir_pattern={args.pathways_gcs_location}"""
-  if args.use_pathways:
-    return yaml.format(args=args)
-  else:
-    return ""
-
-def get_proxy_args(args) -> str:
-  """Arguments for the Pathways proxy.
-  Args:
-    args: user provided arguments for running the command.
-
-  Returns:
-    str: yaml containing arguments for the Pathways proxy.
-  """
-  yaml="""- --alsologtostderr
-              - --v=0
-              - --pathways_ifrt_proxy_server_resource_manager={args.workload}-rm-0-0.{args.workload}:38677
-              - --pathways_ifrt_proxy_server_port=38676
-              - --pathways_tmp_dir_pattern={args.pathways_gcs_location}
-              - --pathways_xprof_trace_enable_bulk_upload=true
-              - --pathways_plaque_network=gcp"""
-  if args.use_pathways:
-    return yaml.format(args=args)
-  else:
-    return ""
-
 def get_system_characteristics(args) -> tuple[SystemCharacteristics|None, int]:
   """Get system characteristics based on user provided arguments.
 
@@ -2894,14 +2803,21 @@ def workload_create(args) -> int:
       )
       xpk_exit(1)
 
+    # Ensure device type is TPUs - currently Pathways supports TPUs only.
+    if system.accelerator_type != AcceleratorType['TPU']:
+      xpk_print(
+          'Currently, Pathways workloads can only be run on TPUs.'
+      )
+      xpk_exit(1)
+
     yml_string = pw_workload_create_yaml.format(args=args,
                                         system=system,
                                         container=container,
                                         accelerator_label=create_accelerator_label(system.accelerator_type, system),
                                         machine_label=create_machine_label(system.accelerator_type, system),
-                                        pathways_rm_args = get_pathways_rm_args(args),
-                                        pathways_worker_args = get_pathways_worker_args(args),
-                                        proxy_args = get_proxy_args(args),
+                                        pathways_rm_args = pw.get_pathways_rm_args(args),
+                                        pathways_worker_args = pw.get_pathways_worker_args(args),
+                                        proxy_args = pw.get_proxy_args(args),
                                         resource_type=resource_type,
                                         local_queue_name=_LOCAL_QUEUE_NAME)
     tmp = write_temporary_file(yml_string)
@@ -3542,9 +3458,7 @@ cluster_create_optional_arguments.add_argument(
 cluster_create_optional_arguments.add_argument(
     '--enable-pathways',
     action='store_true',
-    help=(
-        'Enable cluster to accept Pathways workloads.',
-    ),
+    help='Enable cluster to accept Pathways workloads.',
 )
 cluster_create_optional_arguments.add_argument(
     '--pathways-gce-machine-type',
@@ -3959,7 +3873,7 @@ workload_pathways_workload_arguments.add_argument(
 workload_pathways_workload_arguments.add_argument(
     '--proxy-server-image', 
     type=str,
-    default='gcr.io/cloud-tpu-multipod-dev/pathways-prototype/gke/pathways-demo:proxy_server',
+    default='gcr.io/cloud-tpu-v2-images/pathways/pathways-demo:proxy_server',
     help=(
         'Please provide the proxy server image for Pathways.'
     ),
@@ -3967,7 +3881,7 @@ workload_pathways_workload_arguments.add_argument(
 workload_pathways_workload_arguments.add_argument(
     '--server-image', 
     type=str,
-    default='gcr.io/cloud-tpu-multipod-dev/pathways-prototype/gke/pathways-demo:server',
+    default='gcr.io/cloud-tpu-v2-images/pathways/pathways-demo:server',
     help=(
         'Please provide the server image for Pathways.'
     ),
